+++ project-base/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - PHPStan support (@mhujer)
 	- currently analysing source code by level 0
 - PHP 7.2 support (@TomasLudvik)
+- Uniformity of PHP and Postgres timezones is checked during the build (@Miroslav-Stopka)
 
 ### Changed
 - cache is cleared before PHPUnit tests only when run via [Phing targets](docs/introduction/phing-targets.md), not when run using `phpunit` directly (@PetrHeinz)